Create product-categories-in-query-loop
Add WooCommerce Product Categories to Gutenberg Query Loop as Post Type in order to output querries of Product Categories. This function should replace the need for product-categories shortcode with the flexibility and settings of the Query Loop.

Snippet checks for plugin "Category Order and Taxonomy Terms Order", so that the output order respects ‘orderby’ => ‘term_order’ (the manual sorting defined). If plugin is not installed, sorting falls back to WooCommerce default, alternatively sort by name.  

Snippet loads top level categories when no filters are set. In filters "Parent" is added to select the parent of the categories to show.